+++ Editor/src/com/kotcrab/vis/editor/ui/dialog/NewProjectDialogGeneric.java
@@ -31,7 +31,7 @@ import com.kotcrab.vis.ui.widget.VisLabel;
 import com.kotcrab.vis.ui.widget.VisTable;
 import com.kotcrab.vis.ui.widget.VisTextButton;
 import com.kotcrab.vis.ui.widget.VisValidatableTextField;
-import com.kotcrab.vis.ui.widget.file.FileChooserAdapter;
+import com.kotcrab.vis.ui.widget.file.SingleFileChooserListener;
 
 /**
  * Generic project subdialog for {@link NewProjectDialog}
@@ -107,7 +107,7 @@ public class NewProjectDialogGeneric extends VisTable {
 		chooseRootButton.addListener(new ChangeListener() {
 			@Override
 			public void changed (ChangeEvent event, Actor actor) {
-				fileChooserModule.pickFileOrDirectory(new FileChooserAdapter() {
+				fileChooserModule.pickFileOrDirectory(new SingleFileChooserListener() {
 					@Override
 					public void selected (FileHandle file) {
 						projectRoot.setText(file.file().getAbsolutePath());
@@ -119,7 +119,7 @@ public class NewProjectDialogGeneric extends VisTable {
 		chooseOutputButton.addListener(new ChangeListener() {
 			@Override
 			public void changed (ChangeEvent event, Actor actor) {
-				fileChooserModule.pickFileOrDirectory(new FileChooserAdapter() {
+				fileChooserModule.pickFileOrDirectory(new SingleFileChooserListener() {
 					@Override
 					public void selected (FileHandle file) {
 						outputDirectory.setText(file.file().getAbsolutePath());