+++ UI/src/com/kotcrab/vis/ui/widget/file/FileChooserListener.java
@@ -24,16 +24,9 @@ import com.badlogic.gdx.utils.Array;
  * @author Kotcrab
  */
 public interface FileChooserListener {
-	/** Called when user finished selecting files. */
+	/** Called when user finished selecting files. It is guaranteed that array will contain at least one file. */
 	void selected (Array<FileHandle> files);
 
-	/**
-	 * Called when user finished selecting files. Note that if you are using chooser with multi selection enabled
-	 * you should use {@link #selected(Array)} because this will be invoked only for first file selected.
-	 * @param file file selected by user
-	 */
-	void selected (FileHandle file);
-
-	/** Called when selection dialog was canceled by user */
+	/** Called when selection dialog was canceled by user. */
 	void canceled ();
 }